requirements.txt
pyyaml==5.2
versioneer==0.18
mako==1.1.0
qds-sdk==1.13.2
termcolor
pytest
gitpython
docker
